server/server(Vigil)/requirements.txt
Package                                  Version     Editable project location
---------------------------------------- ----------- -------------------------------------------------------------------------------------------------------------------
aiohttp                                  3.9.3
aiosignal                                1.3.1
altair                                   5.2.0
annotated-types                          0.6.0
anyio                                    4.3.0
async-timeout                            4.0.3
attrs                                    23.2.0
backoff                                  2.2.1
bcrypt                                   4.1.2
blinker                                  1.7.0
cachetools                               5.3.3
certifi                                  2024.2.2
charset-normalizer                       3.3.2
chroma-hnswlib                           0.7.3
chromadb                                 0.4.17
click                                    8.1.7
colorama                                 0.4.6
coloredlogs                              15.0.1
configparser                             5.3.0
dataclasses-json                         0.6.4
datasets                                 2.15.0
Deprecated                               1.2.14
dill                                     0.3.7
distro                                   1.9.0
exceptiongroup                           1.2.0
fastapi                                  0.110.0
filelock                                 3.13.1
Flask                                    3.0.0
flatbuffers                              23.5.26
frozenlist                               1.4.1
fsspec                                   2023.10.0
gitdb                                    4.0.11
GitPython                                3.1.42
google-auth                              2.28.1
googleapis-common-protos                 1.62.0
greenlet                                 3.0.3
grpcio                                   1.62.0
h11                                      0.14.0
httpcore                                 1.0.5
httptools                                0.6.1
httpx                                    0.27.0
huggingface-hub                          0.21.3
humanfriendly                            10.0
idna                                     3.6
importlib-metadata                       6.11.0
importlib_resources                      6.1.2
itsdangerous                             2.1.2
Jinja2                                   3.1.3
joblib                                   1.3.2
jsonpatch                                1.33
jsonpointer                              2.4
jsonschema                               4.21.1
jsonschema-specifications                2023.12.1
kubernetes                               29.0.0
langchain                                0.1.5
langchain-community                      0.0.17
langchain-core                           0.1.18
langchain-openai                         0.0.5
langsmith                                0.0.92
loguru                                   0.7.2
markdown-it-py                           3.0.0
MarkupSafe                               2.1.5
marshmallow                              3.21.1
mdurl                                    0.1.2
monotonic                                1.6
mpmath                                   1.3.0
multidict                                6.0.5
multiprocess                             0.70.15
mypy-extensions                          1.0.0
networkx                                 3.2.1
nltk                                     3.8.1
numpy                                    1.25.2
oauthlib                                 3.2.2
onnxruntime                              1.17.1
openai                                   1.11.1
opentelemetry-api                        1.23.0
opentelemetry-exporter-otlp-proto-common 1.23.0
opentelemetry-exporter-otlp-proto-grpc   1.23.0
opentelemetry-proto                      1.23.0
opentelemetry-sdk                        1.23.0
opentelemetry-semantic-conventions       0.44b0
overrides                                7.7.0
packaging                                23.2
pandas                                   2.0.0
Pillow                                   9.5.0
pip                                      22.3.1
posthog                                  3.5.0
protobuf                                 4.25.3
pulsar-client                            3.4.0
pyarrow                                  14.0.1
pyarrow-hotfix                           0.6
pyasn1                                   0.5.1
pyasn1-modules                           0.3.0
pydantic                                 2.6.4
pydantic_core                            2.16.3
pydeck                                   0.8.0
Pygments                                 2.17.2
Pympler                                  1.0.1
PyPika                                   0.48.9
pyreadline3                              3.4.1
python-dateutil                          2.9.0.post0
python-dotenv                            1.0.1
pytz                                     2024.1
pytz-deprecation-shim                    0.1.0.post0
PyYAML                                   6.0.1
referencing                              0.33.0
regex                                    2023.12.25
requests                                 2.31.0
requests-oauthlib                        1.3.1
rich                                     13.7.1
rpds-py                                  0.18.0
rsa                                      4.9
safetensors                              0.4.2
scikit-learn                             1.4.1.post1
scipy                                    1.12.0
sentence-transformers                    2.2.2
sentencepiece                            0.2.0
setuptools                               65.5.0
six                                      1.16.0
smmap                                    5.0.1
sniffio                                  1.3.1
SQLAlchemy                               2.0.28
starlette                                0.36.3
streamlit                                1.26.0
sympy                                    1.12
tenacity                                 8.2.3
threadpoolctl                            3.3.0
tiktoken                                 0.5.2
tokenizers                               0.15.2
toml                                     0.10.2
toolz                                    0.12.1
torch                                    2.2.1
torchvision                              0.17.1
tornado                                  6.4
tqdm                                     4.66.2
transformers                             4.36.0
typer                                    0.9.0
typing_extensions                        4.10.0
typing-inspect                           0.9.0
tzdata                                   2024.1
tzlocal                                  4.3.1
urllib3                                  2.2.1
uvicorn                                  0.27.1
validators                               0.22.0
vigil-llm                                0.8.7       \path\to\vigil_dir\vigil-llm-0.10.3-alpha
watchdog                                 4.0.0
watchfiles                               0.21.0
websocket-client                         1.7.0
websockets                               12.0
Werkzeug                                 3.0.1
win32-setctime                           1.1.0
wrapt                                    1.16.0
xxhash                                   3.4.1
yara-python                              4.3.1
yarl                                     1.9.4
zipp                                     3.17.0
